==> simulation/notebooks/04_hybrid_strategy.ipynb <==
# Part 4 - Hybrid Strategies

In this stage of the project, we will explore hybrid replenishment strategies. Up until this point, we have examined two distinct strategies, Just-in-Time (JIT) and weekly-scheduled deliveries. However, in reality, businesses often adopt a hybrid model, using JIT for high priority or fast-moving products in combination with routine consolidated shipments for other products.

With the data now transformed and loaded into a SQL database, it is easy to filter and recombine data to simulate a hybrid strategy. We can assign a different replenishment strategy to each product class. This allows us to create a strategy that optimizes operational and financial trade-offs with greater precision. 

from sqlalchemy import create_engine
import sqlalchemy
from sqlalchemy.dialects.postgresql import insert
import numpy as np
import pandas as pd
import math

Connect to the local instance of the PostgreSQL instance created in the last notebook:

engine = create_engine("postgresql+psycopg2://my_user:Password@localhost:5432/warehouse_db")

## Strategies

Class A products, with their higher and more consistent demand, are well suited for frequent targeted restocking and benefit most from a JIT strategy. Class C products on the other hand, benefit from weekly top-ups due to the spordic and slower nature of their demand. The two different replenishment strategies can be combined, taking the best suited strategy for each product class.

Finally, Class B products are an intermediatry group and perform similarly under both replenishment strategies. To determine the best strategy for these products, we will try two separate hybrid strategies, varying on the strategy used for Class B:

- Hybrid A - JIT for Classes A and B, Weekly for Class C
- Hybrid B - JIT for Class A and Weekly for Classes B and C

First we need to return the daily-product level data from the database, then we follow the steps in Notebook 2 to aggregate the data into the daily operations.

### Queries

We can query the database with two separate queries that return the daily product data for both strategies:

# Hybrid A strategy -- JIT for Classes A and B, Weekly for Class C

hybrid_a_product_query = """
    SELECT 
        dp.date,
        dp.product_id,
        dp.warehouse_id,
        'Hybrid A' AS strategy,
        dp.demand,
        dp.inbound_units,
        dp.actual_outbound,
        dp.inventory_level,
        dp.unmet_demand,
        dp.stockout_flag
    FROM daily_products dp
    JOIN dim_products p
    ON dp.product_id = p.product_id
    WHERE (p.product_class = 'C' AND dp.strategy = 'weekly')
        OR (p.product_class IN ('A', 'B') AND dp.strategy = 'JIT')
    ORDER BY 1, 2;
"""

# Hybrid B strategy -- JIT for Class A, Weekly for Classes B and C

hybrid_b_product_query = """
    SELECT 
        dp.date,
        dp.product_id,
        dp.warehouse_id,
        'Hybrid B' AS strategy,
        dp.demand,
        dp.inbound_units,
        dp.actual_outbound,
        dp.inventory_level,
        dp.unmet_demand,
        dp.stockout_flag
    FROM daily_products dp
    JOIN dim_products p
    ON dp.product_id = p.product_id
    WHERE (p.product_class IN ('B', 'C') AND dp.strategy = 'weekly')
        OR (p.product_class = 'A' AND dp.strategy = 'JIT')
    ORDER BY 1, 2;
"""

Return the results to two separate data frames:

hybrid_a_products_df = pd.read_sql_query(hybrid_a_product_query, engine)
hybrid_b_products_df = pd.read_sql_query(hybrid_b_product_query, engine)

### Operations

Follow the same workflow from Notebook 2 to aggregate the product level data for the hybrid strategies to daily operations.

hybrid_a_ops_df = hybrid_a_products_df.groupby(["date", "warehouse_id", "strategy"])

hybrid_a_ops_df = hybrid_a_ops_df.agg({
    "inbound_units": "sum",
    "actual_outbound": "sum",
    "inventory_level": "sum",
    "unmet_demand": "sum",
})

hybrid_a_ops_df = hybrid_a_ops_df.rename(columns={
    "actual_outbound": "orders_fulfilled",
    "unmet_demand": "missed_sales"
})

hybrid_a_ops_df = hybrid_a_ops_df.reset_index()

hybrid_b_ops_df = hybrid_b_products_df.groupby(["date", "warehouse_id", "strategy"])

hybrid_b_ops_df = hybrid_b_ops_df.agg({
    "inbound_units": "sum",
    "actual_outbound": "sum",
    "inventory_level": "sum",
    "unmet_demand": "sum",
})

hybrid_b_ops_df = hybrid_b_ops_df.rename(columns={
    "actual_outbound": "orders_fulfilled",
    "unmet_demand": "missed_sales"
})

hybrid_b_ops_df = hybrid_b_ops_df.reset_index()

#### Outbound

Calculate the number of outbound shipments based on the number of orders fulfilled:

def calculate_outbound_shipments(df, van_capacity=100):
    
    df["outbound_shipments"] = df["orders_fulfilled"] / van_capacity
    df["outbound_shipments"] = df["outbound_shipments"].apply(np.ceil).astype("int")
    df["van_utilization"] = df["orders_fulfilled"] / (df["outbound_shipments"] * van_capacity)

    return df

#### Inbound

Calculate the number of inbound shipments. Because both hybrid strategies replenish Class A products under a JIT strategy, it is unlikely that many single orders will be over 1000 units. Therefore we can use the smaller trucks to replenish stock.

def calculate_inbound_shipments(df, truck_capacity=500):

    df["inbound_shipments"] = df["inbound_units"] / truck_capacity
    df["inbound_shipments"] = df["inbound_shipments"].apply(np.ceil).astype("int")
    df["truck_utilization"] = df["inbound_units"] / (df["inbound_shipments"] * truck_capacity)

    return df

#### Warehouse Utilization

def calculate_warehouse_utilization(df, capacity=2200):
    
    df["warehouse_utilization"] = df["inventory_level"] / capacity

    return df

Apply the functions to both data frames:

hybrid_a_ops_df = calculate_outbound_shipments(hybrid_a_ops_df)
hybrid_a_ops_df = calculate_inbound_shipments(hybrid_a_ops_df)
hybrid_a_ops_df = calculate_warehouse_utilization(hybrid_a_ops_df)

hybrid_b_ops_df = calculate_outbound_shipments(hybrid_b_ops_df)
hybrid_b_ops_df = calculate_inbound_shipments(hybrid_b_ops_df)
hybrid_b_ops_df = calculate_warehouse_utilization(hybrid_b_ops_df)

#### Staff count

Calculate the number of operatives.

def determine_staff_count(row, outbound_column):
    outbound_orders = row[outbound_column]
    workers_needed = math.floor(outbound_orders / 50)
    return min(max(workers_needed, 2), 5)

hybrid_a_ops_df["staff_count"] = hybrid_a_ops_df.apply(determine_staff_count, args=("orders_fulfilled",), axis=1)
hybrid_b_ops_df["staff_count"] = hybrid_b_ops_df.apply(determine_staff_count, args=("orders_fulfilled",), axis=1)

#### Errors

Determine the expected number of errors.

def determine_expected_errors(row, outbound_column, staff_column):
    orders = row[outbound_column]
    staff = row[staff_column]

    if orders == 0 or staff == 0:
        return 0

    base_error_rate = 0.01
    
    orders_per_worker = orders / staff
    error_rate_multiplier = orders_per_worker / 50

    total_error_rate = min(base_error_rate * error_rate_multiplier, 0.05)

    expected_errors = total_error_rate * orders

    return int(expected_errors)


hybrid_a_ops_df["errors"] = hybrid_a_ops_df.apply(determine_expected_errors, args=("orders_fulfilled", "staff_count"), axis=1)
hybrid_b_ops_df["errors"] = hybrid_b_ops_df.apply(determine_expected_errors, args=("orders_fulfilled", "staff_count"), axis=1)

## Update Database

Next we need to write the hybrid strategies data frames to the relevant tables in the database.

The following function writes the data frames to the relevant tables in the database. The <code>conflict_cols</code> variable ensures that the composite primary keys' uniqueness constraint is respected:

def upsert_dataframe(df, table_name, engine, conflict_cols):
    with engine.begin() as conn:
        table = sqlalchemy.Table(table_name, sqlalchemy.MetaData(), autoload_with=engine)
        for _, row in df.iterrows():
            stmt = insert(table).values(row.to_dict())
            stmt = stmt.on_conflict_do_nothing(index_elements=conflict_cols)
            conn.execute(stmt)

Write the Hybrid A tables:

upsert_dataframe(hybrid_a_ops_df, "daily_operations", engine, ["date", "warehouse_id", "strategy"])
upsert_dataframe(hybrid_a_products_df, "daily_products", engine, ["date", "product_id", "warehouse_id", "strategy"])

And the Hybrid B tables:

upsert_dataframe(hybrid_b_ops_df, "daily_operations", engine, ["date", "warehouse_id", "strategy"])
upsert_dataframe(hybrid_b_products_df, "daily_products", engine, ["date", "product_id", "warehouse_id", "strategy"])

### Export Files

Finally, save the hybrid strategies as flat CSV files, so that they may be accessed outside of the database (for example in the GitHub repo).

combined_hybrid_products = pd.concat([hybrid_a_products_df, hybrid_b_products_df], ignore_index=True)
combined_hybrid_operations = pd.concat([hybrid_a_ops_df, hybrid_b_ops_df], ignore_index=True)

combined_hybrid_products.to_csv("../data/warehouse_products_hybrid.csv", index=False)
combined_hybrid_operations.to_csv("../data/warehouse_operations_hybrid.csv", index=False)

In the final notebook, we will explore the data in more detail, comparing the efficiency of each strategy across different metrics.

## Navigation

[Part 3 - Loading Data](03_loading_data.ipynb)

[Part 5 - EDA & Dashboard Prep](05_eda_dashboard_prep.ipynb)
